Moving test functions into test classes.
