+++ tools/make_watertight/tests/mw_unit_test_driver.cc
@@ -1,0 +1,1 @@
+../../../tally/tests/tally_unit_test_driver.cc
